requirements.txt
flask==2.2.5
PyGithub==1.55
paramiko==2.12.0
APScheduler==3.10.1
requests==2.31.0